integration: rocketpool (#390)
* working on scripts, contracts for rpl integration

* sha256 encoder

* testing sha256 combination

* testing sha256 combination

* testing sha256 combination

* testing sha256 combination

* rm encoder

* update contracts to use uint64 for validator fields

* add validator script

* Clean up generation script for rocketpool

* clean up comments

* revert for rpl contract

* rocketpool integration

* add prove validator balance, fix bug related to validator index for combined balance

* add comment for fix

* generate balance proof in rocketpool script

* reorg code

* migrate verify complete validator struct to helper

* migrate verifyValidatorRoot to helper

* migrate all internal fn's to beaconoraclehelper

* rm comment

* mv errors to helper

* fix gitignore

* fix return types

* forge fmt

* add gnosis typechain to turbo build

* add to build folder

* rm unnecessary input for telepathy-relayer build

GitOrigin-RevId: a999ee0bfede7ce77ef9245ed19bb46a8490509f

+++ test/integrations/diva/DivaBeaconOracleTest.t.sol
@@ -23,16 +23,16 @@ struct DepositFixture {
 }
 
 struct StatusFixture {
-    uint256 activationEpoch;
+    uint64 activationEpoch;
     bytes32[] activationProof;
     bytes32 beaconStateRoot;
     bytes32[] beaconStateRootProof;
     bytes32 blockHeaderRoot;
-    uint256 exitEpoch;
+    uint64 exitEpoch;
     bytes32[] exitProof;
-    uint256 pendingEpoch;
+    uint64 pendingEpoch;
     bytes32[] pendingProof;
-    uint256 slashed;
+    uint64 slashed;
     bytes32[] slashedProof;
     uint256 slot;
     uint256 validatorIndex;
@@ -223,7 +223,6 @@ contract DivaBeaconOracleTest is Test, TestEvents, TestErrors {
         assertEq(balance, fixture.validatorBalance);
     }
 
-    // TESTING REVERTS
     function test_RevertInProveDeposit() public {
         vm.expectRevert(
             abi.encodeWithSelector(InvalidDepositProof.selector, depositFixture.pubkeyHash)
@@ -306,7 +305,7 @@ contract DivaBeaconOracleTest is Test, TestEvents, TestErrors {
             beaconStateRootProofInfo,
             validatorProofInfo,
             // Incorrect leaf
-            uint256(0),
+            uint64(0),
             statusFixture.pendingProof,
             BeaconOracleHelper.ValidatorField.ActivationEligibilityEpoch
         );
@@ -322,7 +321,7 @@ contract DivaBeaconOracleTest is Test, TestEvents, TestErrors {
             beaconStateRootProofInfo,
             validatorProofInfo,
             // Incorrect leaf
-            uint256(0),
+            uint64(0),
             statusFixture.activationProof,
             BeaconOracleHelper.ValidatorField.ActivationEpoch
         );
@@ -338,7 +337,7 @@ contract DivaBeaconOracleTest is Test, TestEvents, TestErrors {
             beaconStateRootProofInfo,
             validatorProofInfo,
             // Incorrect leaf
-            uint256(0),
+            uint64(0),
             statusFixture.exitProof,
             BeaconOracleHelper.ValidatorField.ExitEpoch
         );
@@ -353,7 +352,7 @@ contract DivaBeaconOracleTest is Test, TestEvents, TestErrors {
         oracle.proveValidatorField(
             beaconStateRootProofInfo,
             validatorProofInfo, // Validator IS slashed
-            uint256(0),
+            uint64(0),
             statusFixture.slashedProof,
             BeaconOracleHelper.ValidatorField.Slashed
         );
